+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 All notable changes to this project will be documented in this file(see http://keepachangelog.com/).
 
 ## [Unreleased]
+### Added
+- date: support add timezone string on time method.
+
 ### Changed
-- time: set utc to false by default, use current timezone is more reasonable.
+- date: set utc to false by default, use current timezone is more reasonable.
 
 ### Fixed
 - fs: resize string if user provide a large length to read.